Refactor to a more abstract Node type, and separate non-abstract classes into separate files.
